# frontend/static/swagger.json
{
  "swagger": "2.0",
    "info": {
        "title": "User Management API",
        "description": "API for managing users, authentication, and sessions.",
        "version": "1.0.0"
    },
    "host": "localhost:5000",
    "basePath": "/",
    "schemes": ["http"],
    "paths": {
        "/setup_admin": {
            "post": {
                "summary": "Setup the admin account",
                "description": "Sets the initial password for the default admin account.",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "password": {"type": "string"}
                            },
                            "required": ["password"]
                        }
                    }
                ],
                "responses": {
                    "200": {"description": "Admin password set successfully."},
                    "400": {"description": "Admin account does not exist or already set up."}
                }
            }
        },
        "/add_user": {
            "post": {
                "summary": "Add a new user",
                "description": "Adds a new user to the system. Only accessible to admin.",
                "parameters": [
                    {
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "type": "string",
                        "description": "Admin token"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "username": {"type": "string"},
                                "password": {"type": "string"}
                            },
                            "required": ["username", "password"]
                        }
                    }
                ],
                "responses": {
                    "201": {"description": "User added successfully."},
                    "401": {"description": "Invalid or missing admin token."},
                    "400": {"description": "Username already exists or invalid input."}
                }
            }
        },
        "/generate_token": {
            "post": {
                "summary": "Generate authentication token",
                "description": "Generates a JWT token for a user.",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "username": {"type": "string"},
                                "password": {"type": "string"}
                            },
                            "required": ["username", "password"]
                        }
                    }
                ],
                "responses": {
                    "200": {"description": "Token generated successfully."},
                    "401": {"description": "Invalid username or password."}
                }
            }
        },
        "/login": {
            "post": {
                "summary": "User login",
                "description": "Logs in a user and starts a session.",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "username": {"type": "string"},
                                "password": {"type": "string"}
                            },
                            "required": ["username", "password"]
                        }
                    }
                ],
                "responses": {
                    "200": {"description": "Login successful."},
                    "401": {"description": "Invalid username or password."}
                }
            }
        },
        "/logout": {
            "post": {
                "summary": "User logout",
                "description": "Logs out the current user and ends the session.",
                "responses": {
                    "200": {"description": "Logout successful."}
                }
            }
        },
        "/check_session": {
            "get": {
                "summary": "Check active session",
                "description": "Checks if a user is currently logged in.",
                "responses": {
                    "200": {"description": "User is logged in."},
                    "401": {"description": "No active session."}
                }
            }
        }
    }
}